chore: clean up some folder structure
