+++ src/types/mod.rs
@@ -5,8 +5,6 @@ pub use account::*;
 mod action;
 pub use action::*;
 
-pub mod capabilities;
-
 mod coin;
 pub use coin::*;
