+++ tests/e2e/cases/calls.rs
@@ -14,10 +14,9 @@ use relay::{
     rpc::RelayApiClient,
     signers::Eip712PayLoadSigner,
     types::{
-        Call, KeyType, KeyWith712Signer, PrepareCallsCapabilities, PrepareCallsParameters,
-        PrepareCallsResponse, SendPreparedCallsParameters, SendPreparedCallsResponse,
-        SendPreparedCallsSignature, Signature,
-        capabilities::{AuthorizeKey, Meta},
+        AuthorizeKey, Call, KeyType, KeyWith712Signer, Meta, PrepareCallsCapabilities,
+        PrepareCallsParameters, PrepareCallsResponse, SendPreparedCallsParameters,
+        SendPreparedCallsResponse, SendPreparedCallsSignature, Signature,
     },
 };
 use std::str::FromStr;
